Handle message status (#109)
* Update some class to internal

* Temp update seen

* Temp seenBy

* just call seenby one time

* just call with messege not me

* add util func check read message

* update seenUser

* add isSeen to message

* update call api with local

* Fix bug first message

* Update contact loader

* Add some UI not seen and seen

* =)) update token before logout

* finish handle logout

* Temp create socket token

* Stream from application

* Temp

* Fix cannot refresh token

* Finish status

* Finish refresh token

* Format ktlint

* Finish refresh token

* Revert code

+++ core/src/commonMain/kotlin/com/noljanolja/core/user/data/repository/UserRepositoryImpl.kt
@@ -12,7 +12,7 @@ import io.ktor.client.plugins.*
 import io.ktor.client.plugins.auth.*
 import io.ktor.client.plugins.auth.providers.*
 
-class UserRepositoryImpl(
+internal class UserRepositoryImpl(
     private val userRemoteDataSource: UserRemoteDataSource,
     private val authDataSource: AuthDataSource,
     private val client: HttpClient,
@@ -91,15 +91,20 @@ class UserRepositoryImpl(
 
     // Logout
     override suspend fun logout(): Result<Boolean> {
-        return authDataSource.logout().also {
-            val provider =
-                client.plugin(Auth).providers.filterIsInstance<BearerAuthProvider>().firstOrNull()
-            provider?.clearToken()
-            authRepository.delete()
-            localConversationDataSource.deleteAll()
-            localConversationDataSource.deleteAllMessages()
-            localUserDataSource.deleteAllParticipants()
-            localUserDataSource.deleteAll()
+        return pushTokens("").also {
+            if (it.isSuccess) {
+                authDataSource.logout().also {
+                    val provider =
+                        client.plugin(Auth).providers.filterIsInstance<BearerAuthProvider>()
+                            .firstOrNull()
+                    provider?.clearToken()
+                    authRepository.saveAuthToken("")
+                    localConversationDataSource.deleteAll()
+                    localConversationDataSource.deleteAllMessages()
+                    localUserDataSource.deleteAllParticipants()
+                    localUserDataSource.deleteAll()
+                }
+            }
         }
     }
